update to dataflow analysis
The Dataflow class now expects the dataflow analysis
class to model basic blocks rather than individual
instructions. This is useful for some analyses
(e.g., dominators) where the instructions are
not significant. A ModelInstructionsInBasicBlock
helper class is provided for analyses to use as
a default for modeling each instruction in a basic
block in analysis order.