major overhaul of semantic analysis starter code
The driver prints the name of each symbol table.

The starter code is set up to make it a bit more
straightforward to adopt the approach that each
function has one canonical symbol table, created
at the point of the first declaration or definition
of the function. A second declaration or definition
will check the return type and parameter types
against the original.